Add documentation, addresses to gas measurements
Add gas measurement README.
Add BLS Wallet & EOA signer addresses to results.
Split gas measurement index into run & measurement files.
Minor code cleanup & bug fixes.
