+++ .github/workflows/contracts.yml
@@ -33,6 +33,7 @@ jobs:
     - uses: ./.github/actions/setup-contracts-clients
     - run: yarn test
 
+  # ensure gas measurement script runs
   test-gas-measurements:
     runs-on: ubuntu-latest
 
@@ -40,4 +41,4 @@ jobs:
     - uses: actions/checkout@v3
     - uses: ./.github/actions/setup-contracts-clients
     - uses: ./.github/actions/local-contract-deploy
-    - run: yarn hardhat run ./scripts/measure_gas --network gethDev
+    - run: yarn hardhat run ./scripts/measure_gas/run.ts --network gethDev
